added script to run init_database.py file ( bug)
