+++ .env
@@ -1,4 +1,4 @@
-POSTGRES_USER=43df3fa
+POSTGRES_USER=user123
 POSTGRES_PASSWORD=rer23!ds
 POSTGRES_DB=SADatabase
 POSTGRES_HOST=db
