+++ api/database/Dockerfile
@@ -1,0 +1,3 @@
+FROM postgres:alpine
+COPY ./requirements.txt .
+# COPY *.sql /docker-entrypoint-initdb.d
