+++ api/database/docker-entrypoint-initdb.d/2_copy.sql
@@ -1,2 +1,1 @@
 \copy sadata FROM '/data/SA2_info_for_dashboard.csv' with DELIMITER ',' CSV HEADER
-
