+++ api/database/docker-entrypoint-initdb.d/init_database.sh
@@ -1,0 +1,4 @@
+#!usr/bin/python
+echo "-------------- ---- hi from shel script ----------"
+chmod +x ./init_database.py
+python ./init_database.py
