+++ api/database/requirements.txt
@@ -1,0 +1,2 @@
+python==3-alpine
+psycopg2==2.8.5
