+++ docker-compose.yml
@@ -1,22 +1,25 @@
 version: '3'
 services:
   db:
-    image: "postgres:alpine"
+    build: ./api/database
     ports:
       - "5432:5432"
     environment:
       - POSTGRES_USER=${POSTGRES_USER}
       - POSTGRES_PASSWORD=${POSTGRES_PASSWORD}
       - POSTGRES_DB=${POSTGRES_DB}
+      - POSTGRES_HOST=${POSTGRES_HOST}
+      - POSTGRES_PORT=${POSTGRES_PORT}
     volumes:
       - ./src/data/SA2_info_for_dashboard.csv:/data/SA2_info_for_dashboard.csv
+      - ./src/data/SA_SA2s.geojson:/data/SA_SA2s.geojson
       - ./api/database/docker-entrypoint-initdb.d:/docker-entrypoint-initdb.d
+
   api:
     build: ./api
     ports:
       - "5000:5000"
     volumes:
-      - ./src/data/SA_SA2s.geojson:/data/SA_SA2s.geojson
       - ./api:/app
     environment:
       - FLASK_ENV=${FLASK_ENV}
